Pull NewServiceSecretsManager out of httpstate
A sister PR to https://github.com/pulumi/pulumi/pull/11767. This pulls
as much of NewServiceSecretsManager out of httpstate and to the
secrets/service module.

What's left in httpstate is the DefaultSecretManager method pulling off
the client and stack identifier to pass to NewServiceSecretsManager.
That couldn't be done in secrets/service because that would cause a
dependency chain loop (secrets/service would depend on the
httpstate.Stack, but that would depend on secrets/service).

Arguably the `client.Client` ought to just be a string URL and the
secret manager should make it's own client (like it does in
NewServiceSecretsManagerFromState). But trying to keep each individual
change here small.
