+++ pkg/secrets/service/manager.go
@@ -27,6 +27,7 @@ import (
 	"github.com/pulumi/pulumi/sdk/v3/go/common/diag"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/diag/colors"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource/config"
+	"github.com/pulumi/pulumi/sdk/v3/go/common/tokens"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/contract"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/workspace"
 )
@@ -119,15 +120,43 @@ func (sm *serviceSecretsManager) Encrypter() (config.Encrypter, error) {
 	return sm.crypter, nil
 }
 
-func NewServiceSecretsManager(c *client.Client, id client.StackIdentifier) (secrets.Manager, error) {
+func NewServiceSecretsManager(
+	client *client.Client, id client.StackIdentifier, stackName tokens.Name, configFile string) (secrets.Manager, error) {
+
+	contract.Assertf(stackName != "", "stackName %s", "!= \"\"")
+
+	project, _, err := workspace.DetectProjectStackPath(stackName.Q())
+	if err != nil {
+		return nil, err
+	}
+
+	info, err := workspace.LoadProjectStack(project, configFile)
+	if err != nil {
+		return nil, err
+	}
+
+	// We should only save the ProjectStack at this point IF we have changed the
+	// secrets provider. To change the secrets provider to a serviceSecretsManager
+	// we would need to ensure that there are no remnants of the old secret manager
+	// To remove those remnants, we would set those values to be empty in the project
+	// stack, as per changeProjectStackSecretDetails func.
+	// If we do not check to see if the secrets provider has changed, then we will actually
+	// reload the configuration file to be sorted or an empty {} when creating a stack
+	// this is not the desired behaviour.
+	if changeProjectStackSecretDetails(info) {
+		if err := workspace.SaveProjectStack(stackName.Q(), info); err != nil {
+			return nil, err
+		}
+	}
+
 	return &serviceSecretsManager{
 		state: serviceSecretsManagerState{
-			URL:     c.URL(),
+			URL:     client.URL(),
 			Owner:   id.Owner,
 			Project: id.Project,
 			Stack:   id.Stack,
 		},
-		crypter: newServiceCrypter(c, id),
+		crypter: newServiceCrypter(client, id),
 	}, nil
 }
 
@@ -161,4 +190,30 @@ func NewServiceSecretsManagerFromState(state json.RawMessage) (secrets.Manager, 
 		state:   s,
 		crypter: newServiceCrypter(c, id),
 	}, nil
+}
+
+// A passphrase secrets provider has an encryption salt, therefore, changing
+// from passphrase to serviceSecretsManager requires the encryption salt
+// to be removed.
+// A cloud secrets manager has an encryption key and a secrets provider,
+// therefore, changing from cloud to serviceSecretsManager requires the
+// encryption key and secrets provider to be removed.
+// Regardless of what the current secrets provider is, all of these values
+// need to be empty otherwise `getStackSecretsManager` in crypto.go can
+// potentially return the incorrect secret type for the stack.
+func changeProjectStackSecretDetails(info *workspace.ProjectStack) bool {
+	var requiresSave bool
+	if info.SecretsProvider != "" {
+		info.SecretsProvider = ""
+		requiresSave = true
+	}
+	if info.EncryptedKey != "" {
+		info.EncryptedKey = ""
+		requiresSave = true
+	}
+	if info.EncryptionSalt != "" {
+		info.EncryptionSalt = ""
+		requiresSave = true
+	}
+	return requiresSave
 }
